Splitting up worker pool code
