+++ docker-compose.env
@@ -11,7 +11,7 @@ NETNS_MODE=off
 
 # Scheduling
 THREADS_PER_WORKER=1
-PREWARM_TARGET=1
+PREWARM=1
 MAX_QUEUE_RATIO=1
 MAX_WORKERS_PER_FUNCTION=100
 AFFINITY=0
