+++ k8s/config.yml
@@ -12,7 +12,7 @@ data:
   CGROUP_MODE: "on"
   NETNS_MODE: "off"
   THREADS_PER_WORKER: "80"
-  PREWARM_TARGET: "80"
+  PREWARM: "1"
   MAX_QUEUE_RATIO: "1"
   MAX_WORKERS_PER_FUNCTION: "20"
   AFFINITY: "1"
